+++ core/commons/api_container_launcher_lib/api_container_docker_consts/api_container_docker_consts.go
@@ -5,7 +5,8 @@ const (
 	// dictate how the API container ought to behave
 	SerializedArgsEnvVar = "SERIALIZED_ARGS"
 
-	// The location where the Docker volume for storing enclave data will be mounted on the API container
-	// This COULD possibly vary across launcher API verisons, but we can deal with that bridge when we come to it
-	EnclaveDataVolumeMountpoint = "/kurtosis-enclave-data"
+	// The location where the directory on the Docker host machine for storing enclave data will be bind-mounted
+	//  on the API container
+	// This COULD possibly vary across launcher API verisons, but we can cross that bridge when we come to it
+	EnclaveDataDirMountpoint = "/kurtosis-enclave-data"
 )